added a cycle for testing
